+++ docs/changes.rst
@@ -31,6 +31,7 @@ Released on ? 2019.
 * New check for rendered text dimensions.
 * Added support for subtitle formats.
 * Include overall completion stats for languages.
+* Added reporting at project and global scope.
 
 weblate 3.6.1
 -------------